episod 25 is complete.

+++ app/Http/Controllers/CategoriesController.php
@@ -104,10 +104,18 @@ class CategoriesController extends Controller
      */
     public function destroy(Category $category)
     {
-	$category->delete();
+
+        if ($category->posts->count() > 0 ) {
+            session()->flash('error' , 'Category cannot be deleted because it has some posts.');
+
+            return redirect()->back();
+        }
+
+
+	    $category->delete();
 
         session()->flash('success', 'Category deleted successfully.');
 
-	return redirect(route('categories.index'));
+	    return redirect(route('categories.index'));
     }
 }
