+++ app/Http/Controllers/TagsController.php
@@ -104,10 +104,15 @@ class TagsController extends Controller
      */
     public function destroy(Tag $tag)
     {
-	$tag->delete();
+        if ($tag->posts->count() > 0 ) {
+            session()->flash('error' , 'Tag cannot be deleted. because it is associated to some posts.');
+
+            return redirect()->back();
+        }
+	    $tag->delete();
 
         session()->flash('success', 'Tag deleted successfully.');
 
-	return redirect(route('tags.index'));
+	    return redirect(route('tags.index'));
     }
 }
